refactor: organize client internals
