+++ .github/workflows/ci.yml
@@ -16,6 +16,7 @@ permissions:
 
 env:
   CARGO_TERM_COLOR: always
+  PREK_VERSION: 0.3.10
 
 jobs:
   workflow-lint:
@@ -41,26 +42,20 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
         with:
           components: rustfmt, clippy
-      - uses: taiki-e/install-action@prek
+      - uses: taiki-e/install-action@v2
+        with:
+          tool: prek@${{ env.PREK_VERSION }}
       - uses: Swatinem/rust-cache@v2
         with:
           workspaces: |
             . -> target
             xtask -> target
 
-      - name: Validate prek hooks
-        run: prek validate-config prek.toml
-
       - name: Run pre-commit hook set
-        run: prek run --all-files --stage pre-commit
-
-      - name: Format
-        run: |
-          cargo fmt -- --check
-          cargo fmt --manifest-path xtask/Cargo.toml -- --check
+        run: scripts/validate.sh pre-commit
 
       - name: Clippy
-        run: cargo clippy --all-targets --all-features -- -D warnings
+        run: scripts/validate.sh clippy
 
   commit-policy:
     name: Commit policy
@@ -129,13 +124,15 @@ jobs:
             . -> target
             xtask -> target
 
-      - name: Check all features
+      - name: Check all features on MSRV
+        if: matrix.toolchain == '1.88.0'
         run: cargo check --all-features
 
       - name: Test crate
         run: cargo test --locked
 
       - name: Test xtask
+        if: matrix.toolchain == 'stable'
         run: cargo test --locked --manifest-path xtask/Cargo.toml
 
   package:
@@ -149,7 +146,7 @@ jobs:
       - uses: Swatinem/rust-cache@v2
 
       - name: Verify crate package
-        run: cargo package --locked
+        run: scripts/validate.sh package
 
   feature-powerset:
     name: Feature powerset
@@ -160,7 +157,9 @@ jobs:
           persist-credentials: false
       - uses: dtolnay/rust-toolchain@stable
       - uses: Swatinem/rust-cache@v2
-      - uses: taiki-e/install-action@cargo-hack
+      - uses: taiki-e/install-action@v2
+        with:
+          tool: cargo-hack
 
       - name: Check feature combinations
         run: cargo hack check --feature-powerset --no-dev-deps
@@ -180,22 +179,8 @@ jobs:
             . -> target
             xtask -> target
 
-      - name: Generate OpenAPI and Scalar docs
-        run: cargo run --locked --manifest-path xtask/Cargo.toml -- openapi
-
-      - name: Test xtask
-        run: cargo test --locked --manifest-path xtask/Cargo.toml
-
-      - name: Validate generated docs are committed
-        run: |
-          git diff --exit-code -- \
-            openapi/hinge-api.openapi.json \
-            docs/api/openapi.json \
-            docs/api/index.html \
-            docs/api/versions.json
-
-      - name: Build versioned Pages artifact
-        run: scripts/build-pages.sh "$RUNNER_TEMP/pages"
+      - name: Validate OpenAPI and Scalar docs
+        run: scripts/validate.sh docs
 
   dependency-policy:
     name: cargo-deny (${{ matrix.checks }})
